+++ go/src/github.com/koding/kloud/kloud/controller.go
