+++ go/src/github.com/koding/kloud/kloud/keys.go
