+++ go/src/github.com/koding/kloud/kloud/kloud_test.go
