+++ go/src/github.com/koding/kloud/kloud/locker.go
