+++ go/src/github.com/koding/kloud/kloud/storage.go
