+++ go/src/github.com/koding/kloudctl/command/build.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/koding/kite"
-	"github.com/koding/kloud"
+	"github.com/koding/kloud/kloud"
 	"github.com/mitchellh/cli"
 )
