+++ go/src/github.com/koding/kloudctl/command/event.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/koding/kite"
-	"github.com/koding/kloud"
+	"github.com/koding/kloud/kloud"
 	"github.com/mitchellh/cli"
 )
