+++ go/src/koding/kites/kloud/koding/build.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/dgrijalva/jwt-go"
 	kiteprotocol "github.com/koding/kite/protocol"
-	"github.com/koding/kloud"
+	"github.com/koding/kloud/kloud"
 	"github.com/koding/kloud/machinestate"
 	"github.com/koding/kloud/protocol"
 	"github.com/mitchellh/goamz/ec2"
