+++ go/src/koding/kites/kloud/koding/locker.go
@@ -3,7 +3,7 @@ package koding
 import (
 	"time"
 
-	"github.com/koding/kloud"
+	"github.com/koding/kloud/kloud"
 	"labix.org/v2/mgo"
 	"labix.org/v2/mgo/bson"
 )
