+++ go/src/koding/kites/kloud/main.go
@@ -15,7 +15,7 @@ import (
 	"github.com/koding/kite"
 	kiteconfig "github.com/koding/kite/config"
 	"github.com/koding/kite/protocol"
-	"github.com/koding/kloud"
+	"github.com/koding/kloud/kloud"
 	kloudprotocol "github.com/koding/kloud/protocol"
 	"github.com/koding/logging"
 	"github.com/koding/multiconfig"
